rename dummy data name
