+++ shared/src/commonMain/kotlin/musicapp/MusicAppTheme.kt
@@ -1,6 +1,5 @@
 package musicapp
 
-import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Shapes
@@ -15,26 +14,18 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 
 @Composable
-internal fun MyApplicationTheme(
-    darkTheme: Boolean = isSystemInDarkTheme(), content: @Composable () -> Unit
+internal fun MusicAppTheme(
+    content: @Composable () -> Unit
 ) {
-    val colors = //if (darkTheme) {
-        darkColors(
-            primary = Color(0xFF1D2123),
-            primaryVariant = Color(0xFF3700B3),
-            secondary = Color(0xFFFACD66),
-            surface = Color(0xFF1E1E1E),
-            background = Color(0xFF1E1E1E),
-            onSurface = Color(0xFF1E1E1E),
-            // onSurface = Color.White//Color(0x1E1E1E)
-        )
-//    } else {
-//        lightColors(
-//            primary = Color(0xFF6200EE),
-//            primaryVariant = Color(0xFF3700B3),
-//            secondary = Color(0xFF03DAC5)
-//        )
-//    }
+    val colors = darkColors(
+        primary = Color(0xFF1D2123),
+        primaryVariant = Color(0xFF3700B3),
+        secondary = Color(0xFFFACD66),
+        surface = Color(0xFF1E1E1E),
+        background = Color(0xFF1E1E1E),
+        onSurface = Color(0xFF1E1E1E),
+    )
+
     val typography = Typography(
         body1 = TextStyle(
             fontFamily = FontFamily.Default, fontWeight = FontWeight.Normal, fontSize = 16.sp
