+++ shared/src/commonMain/kotlin/musicapp/sampledata/FeaturedPlaylistResponse.kt
@@ -1,4 +1,4 @@
-package musicapp.dummydata
+package musicapp.sampledata
 
 
 val featurePlaylistResponse = """
